+++ plugins/chat.admin.php
@@ -53,7 +53,7 @@ class PluginChatAdmin extends Plugin {
 		$this->setAuthor('undef.de');
 		$this->setCoAuthors('askuri');
 		$this->setVersion('1.0.0');
-		$this->setBuild('2017-05-10');
+		$this->setBuild('2017-05-16');
 		$this->setCopyright('2014 - 2017 by undef.de');
 		$this->setDescription(new Message('chat.admin', 'plugin_description'));
 
@@ -4129,7 +4129,13 @@ class PluginChatAdmin extends Plugin {
 
 						// rename ID filename to map's name
 						$md5new = md5_file($localfile);
-						$filename = $aseco->slugify($gbx->name);
+						$response = new MXInfoFetcher('TM2', $map->uid, true);
+						if ($response !== null && empty($response->error)) {
+							$filename = $response->name;
+						}
+						else {
+							$filename = $aseco->slugify($gbx->name);
+						}
 						$partialdir = $aseco->plugins['PluginRasp']->mxdir . $sepchar . $filename .'_'. $trkid .'.Map.gbx';
 
 						// insure unique filename by incrementing sequence number,